raspberry_pi/requirements.txt
requests==2.31.0
python-socketio==5.10.0
pyserial==3.5
python-dotenv==1.0.0
